penadt

+++ FriendZoneAnt/src/Coordinates/APITest.java
@@ -68,7 +68,7 @@ public class APITest {
     }
     
     public static boolean distFrom(float lat1, float lng1, float lat2, float lng2, int range) {
-        float max_range = (float) range;
+        double max_range = (double) range;
         double earthRadius = 6371000; //meters
         double dLat = Math.toRadians(lat2-lat1);
         double dLng = Math.toRadians(lng2-lng1);
@@ -76,7 +76,7 @@ public class APITest {
                    Math.cos(Math.toRadians(lat1)) * Math.cos(Math.toRadians(lat2)) *
                    Math.sin(dLng/2) * Math.sin(dLng/2);
         double c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1-a));
-        float dist = (float) (earthRadius * c);
+        double dist = (double) (earthRadius * c);
 
         return dist<=max_range;
     }
